Consolidate Graphics acrossWindows & Unix
