+++ src/System.Drawing.Common/src/System/Drawing/Graphics.Windows.cs
@@ -374,142 +374,6 @@ namespace System.Drawing
         }
 
         /// <summary>
-        /// Draws a line connecting the two specified points.
-        /// </summary>
-        public void DrawLine(Pen pen, PointF pt1, PointF pt2)
-        {
-            DrawLine(pen, pt1.X, pt1.Y, pt2.X, pt2.Y);
-        }
-
-        /// <summary>
-        /// Draws a series of line segments that connect an array of points.
-        /// </summary>
-        public unsafe void DrawLines(Pen pen, PointF[] points)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (PointF* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawLines(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p, points.Length));
-            }
-        }
-
-
-        /// <summary>
-        /// Draws a line connecting the two specified points.
-        /// </summary>
-        public void DrawLine(Pen pen, int x1, int y1, int x2, int y2)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-
-            CheckErrorStatus(Gdip.GdipDrawLineI(new HandleRef(this, NativeGraphics), new HandleRef(pen, pen.NativePen), x1, y1, x2, y2));
-        }
-
-        /// <summary>
-        /// Draws a line connecting the two specified points.
-        /// </summary>
-        public void DrawLine(Pen pen, Point pt1, Point pt2)
-        {
-            DrawLine(pen, pt1.X, pt1.Y, pt2.X, pt2.Y);
-        }
-
-        /// <summary>
-        /// Draws a series of line segments that connect an array of points.
-        /// </summary>
-        public unsafe void DrawLines(Pen pen, Point[] points)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (Point* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawLinesI(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p,
-                    points.Length));
-            }
-        }
-
-        /// <summary>
-        /// Draws an arc from the specified ellipse.
-        /// </summary>
-        public void DrawArc(Pen pen, float x, float y, float width, float height, float startAngle, float sweepAngle)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-
-            CheckErrorStatus(Gdip.GdipDrawArc(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(pen, pen.NativePen),
-                x, y, width, height,
-                startAngle,
-                sweepAngle));
-        }
-
-        /// <summary>
-        /// Draws an arc from the specified ellipse.
-        /// </summary>
-        public void DrawArc(Pen pen, RectangleF rect, float startAngle, float sweepAngle)
-        {
-            DrawArc(pen, rect.X, rect.Y, rect.Width, rect.Height, startAngle, sweepAngle);
-        }
-
-        /// <summary>
-        /// Draws an arc from the specified ellipse.
-        /// </summary>
-        public void DrawArc(Pen pen, int x, int y, int width, int height, int startAngle, int sweepAngle)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-
-            CheckErrorStatus(Gdip.GdipDrawArcI(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(pen, pen.NativePen),
-                x, y, width, height,
-                startAngle,
-                sweepAngle));
-        }
-
-        /// <summary>
-        /// Draws an arc from the specified ellipse.
-        /// </summary>
-        public void DrawArc(Pen pen, Rectangle rect, float startAngle, float sweepAngle)
-        {
-            DrawArc(pen, rect.X, rect.Y, rect.Width, rect.Height, startAngle, sweepAngle);
-        }
-
-        /// <summary>
-        /// Draws a cubic bezier curve defined by four ordered pairs that represent points.
-        /// </summary>
-        public void DrawBezier(Pen pen, float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-
-            CheckErrorStatus(Gdip.GdipDrawBezier(
-                new HandleRef(this, NativeGraphics), new HandleRef(pen, pen.NativePen),
-                x1, y1, x2, y2, x3, y3, x4, y4));
-        }
-
-        /// <summary>
-        /// Draws a cubic bezier curve defined by four points.
-        /// </summary>
-        public void DrawBezier(Pen pen, PointF pt1, PointF pt2, PointF pt3, PointF pt4)
-        {
-            DrawBezier(pen, pt1.X, pt1.Y, pt2.X, pt2.Y, pt3.X, pt3.Y, pt4.X, pt4.Y);
-        }
-
-        /// <summary>
         /// Draws a series of cubic Bezier curves from an array of points.
         /// </summary>
         public unsafe void DrawBeziers(Pen pen, PointF[] points)
@@ -529,684 +393,6 @@ namespace System.Drawing
         }
 
         /// <summary>
-        /// Draws a cubic bezier curve defined by four points.
-        /// </summary>
-        public void DrawBezier(Pen pen, Point pt1, Point pt2, Point pt3, Point pt4)
-        {
-            DrawBezier(pen, pt1.X, pt1.Y, pt2.X, pt2.Y, pt3.X, pt3.Y, pt4.X, pt4.Y);
-        }
-
-        /// <summary>
-        /// Draws a series of cubic Bezier curves from an array of points.
-        /// </summary>
-        public unsafe void DrawBeziers(Pen pen, Point[] points)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (Point* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawBeziersI(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p,
-                    points.Length));
-            }
-        }
-
-        /// <summary>
-        /// Draws the outline of a rectangle specified by <paramref name="rect"/>.
-        /// </summary>
-        public void DrawRectangle(Pen pen, Rectangle rect)
-        {
-            DrawRectangle(pen, rect.X, rect.Y, rect.Width, rect.Height);
-        }
-
-        /// <summary>
-        /// Draws the outline of the specified rectangle.
-        /// </summary>
-        public void DrawRectangle(Pen pen, float x, float y, float width, float height)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-
-            CheckErrorStatus(Gdip.GdipDrawRectangle(
-                new HandleRef(this, NativeGraphics), new HandleRef(pen, pen.NativePen),
-                x, y, width, height));
-        }
-
-        /// <summary>
-        /// Draws the outline of the specified rectangle.
-        /// </summary>
-        public void DrawRectangle(Pen pen, int x, int y, int width, int height)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-
-            CheckErrorStatus(Gdip.GdipDrawRectangleI(
-                new HandleRef(this, NativeGraphics), new HandleRef(pen, pen.NativePen),
-                x, y, width, height));
-        }
-
-        /// <summary>
-        /// Draws the outlines of a series of rectangles.
-        /// </summary>
-        public unsafe void DrawRectangles(Pen pen, RectangleF[] rects)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (rects == null)
-                throw new ArgumentNullException(nameof(rects));
-
-            fixed (RectangleF* r = rects)
-            {
-                CheckErrorStatus(Gdip.GdipDrawRectangles(
-                    new HandleRef(this, NativeGraphics), new HandleRef(pen, pen.NativePen),
-                    r, rects.Length));
-            }
-        }
-
-        /// <summary>
-        /// Draws the outlines of a series of rectangles.
-        /// </summary>
-        public unsafe void DrawRectangles(Pen pen, Rectangle[] rects)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (rects == null)
-                throw new ArgumentNullException(nameof(rects));
-
-            fixed (Rectangle* r = rects)
-            {
-                CheckErrorStatus(Gdip.GdipDrawRectanglesI(
-                    new HandleRef(this, NativeGraphics), new HandleRef(pen, pen.NativePen),
-                    r, rects.Length));
-            }
-
-        }
-
-        /// <summary>
-        /// Draws the outline of an ellipse defined by a bounding rectangle.
-        /// </summary>
-        public void DrawEllipse(Pen pen, RectangleF rect)
-        {
-            DrawEllipse(pen, rect.X, rect.Y, rect.Width, rect.Height);
-        }
-
-        /// <summary>
-        /// Draws the outline of an ellipse defined by a bounding rectangle.
-        /// </summary>
-        public void DrawEllipse(Pen pen, float x, float y, float width, float height)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-
-            CheckErrorStatus(Gdip.GdipDrawEllipse(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(pen, pen.NativePen),
-                x, y, width, height));
-        }
-
-        /// <summary>
-        /// Draws the outline of an ellipse specified by a bounding rectangle.
-        /// </summary>
-        public void DrawEllipse(Pen pen, Rectangle rect)
-        {
-            DrawEllipse(pen, rect.X, rect.Y, rect.Width, rect.Height);
-        }
-
-        /// <summary>
-        /// Draws the outline of an ellipse defined by a bounding rectangle.
-        /// </summary>
-        public void DrawEllipse(Pen pen, int x, int y, int width, int height)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-
-            CheckErrorStatus(Gdip.GdipDrawEllipseI(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(pen, pen.NativePen),
-                x, y, width, height));
-        }
-
-        /// <summary>
-        /// Draws the outline of a pie section defined by an ellipse and two radial lines.
-        /// </summary>
-        public void DrawPie(Pen pen, RectangleF rect, float startAngle, float sweepAngle)
-        {
-            DrawPie(pen, rect.X, rect.Y, rect.Width, rect.Height, startAngle, sweepAngle);
-        }
-
-        /// <summary>
-        /// Draws the outline of a pie section defined by an ellipse and two radial lines.
-        /// </summary>
-        public void DrawPie(Pen pen, float x, float y, float width, float height, float startAngle, float sweepAngle)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-
-            CheckErrorStatus(Gdip.GdipDrawPie(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(pen, pen.NativePen),
-                x, y, width, height,
-                startAngle,
-                sweepAngle));
-        }
-
-        /// <summary>
-        /// Draws the outline of a pie section defined by an ellipse and two radial lines.
-        /// </summary>
-        public void DrawPie(Pen pen, Rectangle rect, float startAngle, float sweepAngle)
-        {
-            DrawPie(pen, rect.X, rect.Y, rect.Width, rect.Height, startAngle, sweepAngle);
-        }
-
-        /// <summary>
-        /// Draws the outline of a pie section defined by an ellipse and two radial lines.
-        /// </summary>
-        public void DrawPie(Pen pen, int x, int y, int width, int height, int startAngle, int sweepAngle)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-
-            CheckErrorStatus(Gdip.GdipDrawPieI(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(pen, pen.NativePen),
-                x, y, width, height,
-                startAngle,
-                sweepAngle));
-        }
-
-        /// <summary>
-        /// Draws the outline of a polygon defined by an array of points.
-        /// </summary>
-        public unsafe void DrawPolygon(Pen pen, PointF[] points)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (PointF* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawPolygon(
-                    new HandleRef(this, NativeGraphics), new HandleRef(pen, pen.NativePen),
-                    p, points.Length));
-            }
-        }
-
-        /// <summary>
-        /// Draws the outline of a polygon defined by an array of points.
-        /// </summary>
-        public unsafe void DrawPolygon(Pen pen, Point[] points)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (Point* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawPolygonI(
-                    new HandleRef(this, NativeGraphics), new HandleRef(pen, pen.NativePen),
-                    p, points.Length));
-            }
-        }
-
-        /// <summary>
-        /// Draws the lines and curves defined by a <see cref='GraphicsPath'/>.
-        /// </summary>
-        public void DrawPath(Pen pen, GraphicsPath path)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (path == null)
-                throw new ArgumentNullException(nameof(path));
-
-            CheckErrorStatus(Gdip.GdipDrawPath(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(pen, pen.NativePen),
-                new HandleRef(path, path._nativePath)));
-        }
-
-        /// <summary>
-        /// Draws a curve defined by an array of points.
-        /// </summary>
-        public unsafe void DrawCurve(Pen pen, PointF[] points)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (PointF* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawCurve(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p, points.Length));
-            }
-        }
-
-        /// <summary>
-        /// Draws a curve defined by an array of points.
-        /// </summary>
-        public unsafe void DrawCurve(Pen pen, PointF[] points, float tension)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (PointF* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawCurve2(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p, points.Length,
-                    tension));
-            }
-        }
-
-        public void DrawCurve(Pen pen, PointF[] points, int offset, int numberOfSegments)
-        {
-            DrawCurve(pen, points, offset, numberOfSegments, 0.5f);
-        }
-
-        /// <summary>
-        /// Draws a curve defined by an array of points.
-        /// </summary>
-        public unsafe void DrawCurve(Pen pen, PointF[] points, int offset, int numberOfSegments, float tension)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (PointF* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawCurve3(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p, points.Length,
-                    offset,
-                    numberOfSegments,
-                    tension));
-            }
-        }
-
-        /// <summary>
-        /// Draws a curve defined by an array of points.
-        /// </summary>
-        public unsafe void DrawCurve(Pen pen, Point[] points)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (Point* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawCurveI(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p, points.Length));
-            }
-        }
-
-        /// <summary>
-        /// Draws a curve defined by an array of points.
-        /// </summary>
-        public unsafe void DrawCurve(Pen pen, Point[] points, float tension)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (Point* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawCurve2I(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p, points.Length,
-                    tension));
-            }
-        }
-
-        /// <summary>
-        /// Draws a curve defined by an array of points.
-        /// </summary>
-        public unsafe void DrawCurve(Pen pen, Point[] points, int offset, int numberOfSegments, float tension)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (Point* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawCurve3I(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p, points.Length,
-                    offset,
-                    numberOfSegments,
-                    tension));
-            }
-        }
-
-        /// <summary>
-        /// Draws a closed curve defined by an array of points.
-        /// </summary>
-        public unsafe void DrawClosedCurve(Pen pen, PointF[] points)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (PointF* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawClosedCurve(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p, points.Length));
-            }
-        }
-
-        /// <summary>
-        /// Draws a closed curve defined by an array of points.
-        /// </summary>
-        public unsafe void DrawClosedCurve(Pen pen, PointF[] points, float tension, FillMode fillmode)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (PointF* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawClosedCurve2(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p, points.Length,
-                    tension));
-            }
-        }
-
-        /// <summary>
-        /// Draws a closed curve defined by an array of points.
-        /// </summary>
-        public unsafe void DrawClosedCurve(Pen pen, Point[] points)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (Point* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawClosedCurveI(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p, points.Length));
-            }
-        }
-
-        /// <summary>
-        /// Draws a closed curve defined by an array of points.
-        /// </summary>
-        public unsafe void DrawClosedCurve(Pen pen, Point[] points, float tension, FillMode fillmode)
-        {
-            if (pen == null)
-                throw new ArgumentNullException(nameof(pen));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (Point* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipDrawClosedCurve2I(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(pen, pen.NativePen),
-                    p, points.Length,
-                    tension));
-            }
-        }
-
-        /// <summary>
-        /// Fills the entire drawing surface with the specified color.
-        /// </summary>
-        public void Clear(Color color)
-        {
-            Gdip.CheckStatus(Gdip.GdipGraphicsClear(new HandleRef(this, NativeGraphics), color.ToArgb()));
-        }
-
-        /// <summary>
-        /// Fills the interior of a rectangle with a <see cref='Brush'/>.
-        /// </summary>
-        public void FillRectangle(Brush brush, RectangleF rect)
-        {
-            FillRectangle(brush, rect.X, rect.Y, rect.Width, rect.Height);
-        }
-
-        /// <summary>
-        /// Fills the interior of a rectangle with a <see cref='Brush'/>.
-        /// </summary>
-        public void FillRectangle(Brush brush, float x, float y, float width, float height)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-
-            CheckErrorStatus(Gdip.GdipFillRectangle(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(brush, brush.NativeBrush),
-                x, y, width, height));
-        }
-
-        /// <summary>
-        /// Fills the interior of a rectangle with a <see cref='Brush'/>.
-        /// </summary>
-        public void FillRectangle(Brush brush, Rectangle rect)
-        {
-            FillRectangle(brush, rect.X, rect.Y, rect.Width, rect.Height);
-        }
-
-        /// <summary>
-        /// Fills the interior of a rectangle with a <see cref='Brush'/>.
-        /// </summary>
-        public void FillRectangle(Brush brush, int x, int y, int width, int height)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-
-            CheckErrorStatus(Gdip.GdipFillRectangleI(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(brush, brush.NativeBrush),
-                x, y, width, height));
-        }
-
-        /// <summary>
-        /// Fills the interiors of a series of rectangles with a <see cref='Brush'/>.
-        /// </summary>
-        public unsafe void FillRectangles(Brush brush, RectangleF[] rects)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-            if (rects == null)
-                throw new ArgumentNullException(nameof(rects));
-
-            fixed (RectangleF* r = rects)
-            {
-                CheckErrorStatus(Gdip.GdipFillRectangles(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(brush, brush.NativeBrush),
-                    r, rects.Length));
-            }
-        }
-
-        /// <summary>
-        /// Fills the interiors of a series of rectangles with a <see cref='Brush'/>.
-        /// </summary>
-        public unsafe void FillRectangles(Brush brush, Rectangle[] rects)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-            if (rects == null)
-                throw new ArgumentNullException(nameof(rects));
-
-            fixed (Rectangle* r = rects)
-            {
-                CheckErrorStatus(Gdip.GdipFillRectanglesI(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(brush, brush.NativeBrush),
-                    r, rects.Length));
-            }
-        }
-
-        /// <summary>
-        /// Fills the interior of a polygon defined by an array of points.
-        /// </summary>
-        public void FillPolygon(Brush brush, PointF[] points)
-        {
-            FillPolygon(brush, points, FillMode.Alternate);
-        }
-
-        /// <summary>
-        /// Fills the interior of a polygon defined by an array of points.
-        /// </summary>
-        public unsafe void FillPolygon(Brush brush, PointF[] points, FillMode fillMode)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (PointF* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipFillPolygon(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(brush, brush.NativeBrush),
-                    p, points.Length,
-                    fillMode));
-            }
-        }
-
-        /// <summary>
-        /// Fills the interior of a polygon defined by an array of points.
-        /// </summary>
-        public void FillPolygon(Brush brush, Point[] points)
-        {
-            FillPolygon(brush, points, FillMode.Alternate);
-        }
-
-        /// <summary>
-        /// Fills the interior of a polygon defined by an array of points.
-        /// </summary>
-        public unsafe void FillPolygon(Brush brush, Point[] points, FillMode fillMode)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (Point* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipFillPolygonI(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(brush, brush.NativeBrush),
-                    p, points.Length,
-                    fillMode));
-            }
-        }
-
-        /// <summary>
-        /// Fills the interior of an ellipse defined by a bounding rectangle.
-        /// </summary>
-        public void FillEllipse(Brush brush, RectangleF rect)
-        {
-            FillEllipse(brush, rect.X, rect.Y, rect.Width, rect.Height);
-        }
-
-        /// <summary>
-        /// Fills the interior of an ellipse defined by a bounding rectangle.
-        /// </summary>
-        public void FillEllipse(Brush brush, float x, float y, float width, float height)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-
-            CheckErrorStatus(Gdip.GdipFillEllipse(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(brush, brush.NativeBrush),
-                x, y, width, height));
-        }
-
-        /// <summary>
-        /// Fills the interior of an ellipse defined by a bounding rectangle.
-        /// </summary>
-        public void FillEllipse(Brush brush, Rectangle rect)
-        {
-            FillEllipse(brush, rect.X, rect.Y, rect.Width, rect.Height);
-        }
-
-        /// <summary>
-        /// Fills the interior of an ellipse defined by a bounding rectangle.
-        /// </summary>
-        public void FillEllipse(Brush brush, int x, int y, int width, int height)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-
-            CheckErrorStatus(Gdip.GdipFillEllipseI(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(brush, brush.NativeBrush),
-                x, y, width, height));
-        }
-
-        /// <summary>
-        /// Fills the interior of a pie section defined by an ellipse and two radial lines.
-        /// </summary>
-        public void FillPie(Brush brush, Rectangle rect, float startAngle, float sweepAngle)
-        {
-            FillPie(brush, rect.X, rect.Y, rect.Width, rect.Height, startAngle, sweepAngle);
-        }
-
-        /// <summary>
-        /// Fills the interior of a pie section defined by an ellipse and two radial lines.
-        /// </summary>
-        public void FillPie(Brush brush, float x, float y, float width, float height, float startAngle, float sweepAngle)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-
-            CheckErrorStatus(Gdip.GdipFillPie(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(brush, brush.NativeBrush),
-                x, y, width, height,
-                startAngle,
-                sweepAngle));
-        }
-
-        /// <summary>
-        /// Fills the interior of a pie section defined by an ellipse and two radial lines.
-        /// </summary>
-        public void FillPie(Brush brush, int x, int y, int width, int height, int startAngle, int sweepAngle)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-
-            CheckErrorStatus(Gdip.GdipFillPieI(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(brush, brush.NativeBrush),
-                x, y, width, height,
-                startAngle,
-                sweepAngle));
-        }
-
-        /// <summary>
         /// Fills the interior of a path.
         /// </summary>
         public void FillPath(Brush brush, GraphicsPath path)
@@ -1223,93 +409,6 @@ namespace System.Drawing
         }
 
         /// <summary>
-        /// Fills the interior a closed curve defined by an array of points.
-        /// </summary>
-        public unsafe void FillClosedCurve(Brush brush, PointF[] points)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (PointF* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipFillClosedCurve(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(brush, brush.NativeBrush),
-                    p, points.Length));
-            }
-        }
-
-        /// <summary>
-        /// Fills the interior of a closed curve defined by an array of points.
-        /// </summary>
-        public void FillClosedCurve(Brush brush, PointF[] points, FillMode fillmode)
-        {
-            FillClosedCurve(brush, points, fillmode, 0.5f);
-        }
-
-        public unsafe void FillClosedCurve(Brush brush, PointF[] points, FillMode fillmode, float tension)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (PointF* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipFillClosedCurve2(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(brush, brush.NativeBrush),
-                    p, points.Length,
-                    tension,
-                    fillmode));
-            }
-        }
-
-        /// <summary>
-        /// Fills the interior a closed curve defined by an array of points.
-        /// </summary>
-        public unsafe void FillClosedCurve(Brush brush, Point[] points)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (Point* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipFillClosedCurveI(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(brush, brush.NativeBrush),
-                    p, points.Length));
-            }
-        }
-
-        public void FillClosedCurve(Brush brush, Point[] points, FillMode fillmode)
-        {
-            FillClosedCurve(brush, points, fillmode, 0.5f);
-        }
-
-        public unsafe void FillClosedCurve(Brush brush, Point[] points, FillMode fillmode, float tension)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-            if (points == null)
-                throw new ArgumentNullException(nameof(points));
-
-            fixed (Point* p = points)
-            {
-                CheckErrorStatus(Gdip.GdipFillClosedCurve2I(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(brush, brush.NativeBrush),
-                    p, points.Length,
-                    tension,
-                    fillmode));
-            }
-        }
-
-        /// <summary>
         /// Fills the interior of a <see cref='Region'/>.
         /// </summary>
         public void FillRegion(Brush brush, Region region)
@@ -1323,186 +422,6 @@ namespace System.Drawing
                 new HandleRef(this, NativeGraphics),
                 new HandleRef(brush, brush.NativeBrush),
                 new HandleRef(region, region.NativeRegion)));
-        }
-
-        /// <summary>
-        /// Draws a string with the specified font.
-        /// </summary>
-        public void DrawString(string s, Font font, Brush brush, float x, float y)
-        {
-            DrawString(s, font, brush, new RectangleF(x, y, 0, 0), null);
-        }
-
-        public void DrawString(string s, Font font, Brush brush, PointF point)
-        {
-            DrawString(s, font, brush, new RectangleF(point.X, point.Y, 0, 0), null);
-        }
-
-        public void DrawString(string s, Font font, Brush brush, float x, float y, StringFormat format)
-        {
-            DrawString(s, font, brush, new RectangleF(x, y, 0, 0), format);
-        }
-
-        public void DrawString(string s, Font font, Brush brush, PointF point, StringFormat format)
-        {
-            DrawString(s, font, brush, new RectangleF(point.X, point.Y, 0, 0), format);
-        }
-
-        public void DrawString(string s, Font font, Brush brush, RectangleF layoutRectangle)
-        {
-            DrawString(s, font, brush, layoutRectangle, null);
-        }
-
-        public void DrawString(string s, Font font, Brush brush, RectangleF layoutRectangle, StringFormat format)
-        {
-            if (brush == null)
-                throw new ArgumentNullException(nameof(brush));
-            if (string.IsNullOrEmpty(s))
-                return;
-            if (font == null)
-                throw new ArgumentNullException(nameof(font));
-
-            CheckErrorStatus(Gdip.GdipDrawString(
-                new HandleRef(this, NativeGraphics),
-                s,
-                s.Length,
-                new HandleRef(font, font.NativeFont),
-                ref layoutRectangle,
-                new HandleRef(format, format?.nativeFormat ?? IntPtr.Zero),
-                new HandleRef(brush, brush.NativeBrush)));
-        }
-
-        public SizeF MeasureString(
-            string text,
-            Font font,
-            SizeF layoutArea,
-            StringFormat stringFormat,
-            out int charactersFitted,
-            out int linesFilled)
-        {
-            if (string.IsNullOrEmpty(text))
-            {
-                charactersFitted = 0;
-                linesFilled = 0;
-                return SizeF.Empty;
-            }
-
-            if (font == null)
-                throw new ArgumentNullException(nameof(font));
-
-            RectangleF layout = new RectangleF(0, 0, layoutArea.Width, layoutArea.Height);
-            RectangleF boundingBox = new RectangleF();
-
-            Gdip.CheckStatus(Gdip.GdipMeasureString(
-                new HandleRef(this, NativeGraphics),
-                text,
-                text.Length,
-                new HandleRef(font, font.NativeFont),
-                ref layout,
-                new HandleRef(stringFormat, stringFormat?.nativeFormat ?? IntPtr.Zero),
-                ref boundingBox,
-                out charactersFitted,
-                out linesFilled));
-
-            return boundingBox.Size;
-        }
-
-        public SizeF MeasureString(string text, Font font, PointF origin, StringFormat stringFormat)
-        {
-            if (string.IsNullOrEmpty(text))
-                return SizeF.Empty;
-            if (font == null)
-                throw new ArgumentNullException(nameof(font));
-
-            RectangleF layout = new RectangleF(origin.X, origin.Y, 0, 0);
-            RectangleF boundingBox = new RectangleF();
-
-            Gdip.CheckStatus(Gdip.GdipMeasureString(
-                new HandleRef(this, NativeGraphics),
-                text,
-                text.Length,
-                new HandleRef(font, font.NativeFont),
-                ref layout,
-                new HandleRef(stringFormat, stringFormat?.nativeFormat ?? IntPtr.Zero),
-                ref boundingBox,
-                out int a,
-                out int b));
-
-            return boundingBox.Size;
-        }
-
-        public SizeF MeasureString(string text, Font font, SizeF layoutArea) => MeasureString(text, font, layoutArea, null);
-
-        public SizeF MeasureString(string text, Font font, SizeF layoutArea, StringFormat stringFormat)
-        {
-            if (string.IsNullOrEmpty(text))
-                return SizeF.Empty;
-            if (font == null)
-                throw new ArgumentNullException(nameof(font));
-
-            RectangleF layout = new RectangleF(0, 0, layoutArea.Width, layoutArea.Height);
-            RectangleF boundingBox = new RectangleF();
-
-            Gdip.CheckStatus(Gdip.GdipMeasureString(
-                new HandleRef(this, NativeGraphics),
-                text,
-                text.Length,
-                new HandleRef(font, font.NativeFont),
-                ref layout,
-                new HandleRef(stringFormat, stringFormat?.nativeFormat ?? IntPtr.Zero),
-                ref boundingBox,
-                out int a,
-                out int b));
-
-            return boundingBox.Size;
-        }
-
-        public SizeF MeasureString(string text, Font font)
-        {
-            return MeasureString(text, font, new SizeF(0, 0));
-        }
-
-        public SizeF MeasureString(string text, Font font, int width)
-        {
-            return MeasureString(text, font, new SizeF(width, 999999));
-        }
-
-        public SizeF MeasureString(string text, Font font, int width, StringFormat format)
-        {
-            return MeasureString(text, font, new SizeF(width, 999999), format);
-        }
-
-        public Region[] MeasureCharacterRanges(string text, Font font, RectangleF layoutRect, StringFormat stringFormat)
-        {
-            if (string.IsNullOrEmpty(text))
-                return Array.Empty<Region>();
-            if (font == null)
-                throw new ArgumentNullException(nameof(font));
-
-            Gdip.CheckStatus(Gdip.GdipGetStringFormatMeasurableCharacterRangeCount(
-                new HandleRef(stringFormat, stringFormat?.nativeFormat ?? IntPtr.Zero),
-                out int count));
-
-            IntPtr[] gpRegions = new IntPtr[count];
-            Region[] regions = new Region[count];
-
-            for (int f = 0; f < count; f++)
-            {
-                regions[f] = new Region();
-                gpRegions[f] = regions[f].NativeRegion;
-            }
-
-            Gdip.CheckStatus(Gdip.GdipMeasureCharacterRanges(
-                new HandleRef(this, NativeGraphics),
-                text,
-                text.Length,
-                new HandleRef(font, font.NativeFont),
-                ref layoutRect,
-                new HandleRef(stringFormat, stringFormat?.nativeFormat ?? IntPtr.Zero),
-                count,
-                gpRegions));
-
-            return regions;
         }
 
         public void DrawIcon(Icon icon, int x, int y)
@@ -1566,552 +485,6 @@ namespace System.Drawing
             {
                 icon.DrawUnstretched(this, targetRect);
             }
-        }
-
-        /// <summary>
-        /// Draws the specified image at the specified location.
-        /// </summary>
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, PointF point)
-        {
-            DrawImage(image, point.X, point.Y);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, float x, float y)
-        {
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int status = Gdip.GdipDrawImage(
-                new HandleRef(this, NativeGraphics), new HandleRef(image, image.nativeImage),
-                x, y);
-
-            IgnoreMetafileErrors(image, ref status);
-            CheckErrorStatus(status);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, RectangleF rect)
-        {
-            DrawImage(image, rect.X, rect.Y, rect.Width, rect.Height);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, float x, float y, float width, float height)
-        {
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int status = Gdip.GdipDrawImageRect(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(image, image.nativeImage),
-                x, y,
-                width, height);
-
-            IgnoreMetafileErrors(image, ref status);
-            CheckErrorStatus(status);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, Point point)
-        {
-            DrawImage(image, point.X, point.Y);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, int x, int y)
-        {
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int status = Gdip.GdipDrawImageI(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(image, image.nativeImage),
-                x, y);
-
-            IgnoreMetafileErrors(image, ref status);
-            CheckErrorStatus(status);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, Rectangle rect)
-        {
-            DrawImage(image, rect.X, rect.Y, rect.Width, rect.Height);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, int x, int y, int width, int height)
-        {
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int status = Gdip.GdipDrawImageRectI(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(image, image.nativeImage),
-                x, y,
-                width, height);
-
-            IgnoreMetafileErrors(image, ref status);
-            CheckErrorStatus(status);
-        }
-
-        public void DrawImageUnscaled(Image image, Point point)
-        {
-            DrawImage(image, point.X, point.Y);
-        }
-
-        public void DrawImageUnscaled(Image image, int x, int y)
-        {
-            DrawImage(image, x, y);
-        }
-
-        public void DrawImageUnscaled(Image image, Rectangle rect)
-        {
-            DrawImage(image, rect.X, rect.Y);
-        }
-
-        public void DrawImageUnscaled(Image image, int x, int y, int width, int height)
-        {
-            DrawImage(image, x, y);
-        }
-
-        public void DrawImageUnscaledAndClipped(Image image, Rectangle rect)
-        {
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int width = Math.Min(rect.Width, image.Width);
-            int height = Math.Min(rect.Height, image.Height);
-
-            // We could put centering logic here too for the case when the image
-            // is smaller than the rect.
-            DrawImage(image, rect, 0, 0, width, height, GraphicsUnit.Pixel);
-        }
-
-        // Affine or perspective blt
-        //  destPoints.Length = 3: rect => parallelogram
-        // destPoints[0] <=> top-left corner of the source rectangle
-        //      destPoints[1] <=> top-right corner
-        //       destPoints[2] <=> bottom-left corner
-        //  destPoints.Length = 4: rect => quad
-        // destPoints[3] <=> bottom-right corner
-        //
-        //  @notes Perspective blt only works for bitmap images.
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public unsafe void DrawImage(Image image, PointF[] destPoints)
-        {
-            if (destPoints == null)
-                throw new ArgumentNullException(nameof(destPoints));
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int count = destPoints.Length;
-            if (count != 3 && count != 4)
-                throw new ArgumentException(SR.GdiplusDestPointsInvalidLength);
-
-            fixed (PointF* p = destPoints)
-            {
-                int status = Gdip.GdipDrawImagePoints(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(image, image.nativeImage),
-                    p, count);
-
-                IgnoreMetafileErrors(image, ref status);
-                CheckErrorStatus(status);
-            }
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public unsafe void DrawImage(Image image, Point[] destPoints)
-        {
-            if (destPoints == null)
-                throw new ArgumentNullException(nameof(destPoints));
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int count = destPoints.Length;
-            if (count != 3 && count != 4)
-                throw new ArgumentException(SR.GdiplusDestPointsInvalidLength);
-
-            fixed (Point* p = destPoints)
-            {
-                int status = Gdip.GdipDrawImagePointsI(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(image, image.nativeImage),
-                    p, count);
-
-                IgnoreMetafileErrors(image, ref status);
-                CheckErrorStatus(status);
-            }
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, float x, float y, RectangleF srcRect, GraphicsUnit srcUnit)
-        {
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int status = Gdip.GdipDrawImagePointRect(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(image, image.nativeImage),
-                x, y,
-                srcRect.X, srcRect.Y, srcRect.Width, srcRect.Height,
-                (int)srcUnit);
-
-            IgnoreMetafileErrors(image, ref status);
-            CheckErrorStatus(status);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, int x, int y, Rectangle srcRect, GraphicsUnit srcUnit)
-        {
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int status = Gdip.GdipDrawImagePointRectI(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(image, image.nativeImage),
-                x, y,
-                srcRect.X, srcRect.Y, srcRect.Width, srcRect.Height,
-                (int)srcUnit);
-
-            IgnoreMetafileErrors(image, ref status);
-            CheckErrorStatus(status);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, RectangleF destRect, RectangleF srcRect, GraphicsUnit srcUnit)
-        {
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int status = Gdip.GdipDrawImageRectRect(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(image, image.nativeImage),
-                destRect.X, destRect.Y, destRect.Width, destRect.Height,
-                srcRect.X, srcRect.Y, srcRect.Width, srcRect.Height,
-                srcUnit,
-                NativeMethods.NullHandleRef,
-                null,
-                IntPtr.Zero);
-
-            IgnoreMetafileErrors(image, ref status);
-            CheckErrorStatus(status);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, Rectangle destRect, Rectangle srcRect, GraphicsUnit srcUnit)
-        {
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int status = Gdip.GdipDrawImageRectRectI(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(image, image.nativeImage),
-                destRect.X, destRect.Y, destRect.Width, destRect.Height,
-                srcRect.X, srcRect.Y, srcRect.Width, srcRect.Height,
-                srcUnit,
-                NativeMethods.NullHandleRef,
-                null,
-                IntPtr.Zero);
-
-            IgnoreMetafileErrors(image, ref status);
-            CheckErrorStatus(status);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public unsafe void DrawImage(Image image, PointF[] destPoints, RectangleF srcRect, GraphicsUnit srcUnit)
-        {
-            if (destPoints == null)
-                throw new ArgumentNullException(nameof(destPoints));
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int count = destPoints.Length;
-            if (count != 3 && count != 4)
-                throw new ArgumentException(SR.GdiplusDestPointsInvalidLength);
-
-            fixed (PointF* p = destPoints)
-            {
-                int status = Gdip.GdipDrawImagePointsRect(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(image, image.nativeImage),
-                    p, destPoints.Length,
-                    srcRect.X, srcRect.Y, srcRect.Width, srcRect.Height,
-                    srcUnit,
-                    NativeMethods.NullHandleRef,
-                    null,
-                    IntPtr.Zero);
-
-                IgnoreMetafileErrors(image, ref status);
-                CheckErrorStatus(status);
-            }
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, PointF[] destPoints, RectangleF srcRect, GraphicsUnit srcUnit, ImageAttributes imageAttr)
-        {
-            DrawImage(image, destPoints, srcRect, srcUnit, imageAttr, null, 0);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(
-            Image image,
-            PointF[] destPoints,
-            RectangleF srcRect,
-            GraphicsUnit srcUnit,
-            ImageAttributes imageAttr,
-            DrawImageAbort callback)
-        {
-            DrawImage(image, destPoints, srcRect, srcUnit, imageAttr, callback, 0);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public unsafe void DrawImage(
-            Image image,
-            PointF[] destPoints,
-            RectangleF srcRect,
-            GraphicsUnit srcUnit,
-            ImageAttributes imageAttr,
-            DrawImageAbort callback,
-            int callbackData)
-        {
-            if (destPoints == null)
-                throw new ArgumentNullException(nameof(destPoints));
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int count = destPoints.Length;
-            if (count != 3 && count != 4)
-                throw new ArgumentException(SR.GdiplusDestPointsInvalidLength);
-
-            fixed (PointF* p = destPoints)
-            {
-                int status = Gdip.GdipDrawImagePointsRect(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(image, image.nativeImage),
-                    p, destPoints.Length,
-                    srcRect.X, srcRect.Y, srcRect.Width, srcRect.Height,
-                    srcUnit,
-                    new HandleRef(imageAttr, imageAttr?.nativeImageAttributes ?? IntPtr.Zero),
-                    callback,
-                    (IntPtr)callbackData);
-
-                IgnoreMetafileErrors(image, ref status);
-                CheckErrorStatus(status);
-            }
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(Image image, Point[] destPoints, Rectangle srcRect, GraphicsUnit srcUnit)
-        {
-            DrawImage(image, destPoints, srcRect, srcUnit, null, null, 0);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(
-            Image image,
-            Point[] destPoints,
-            Rectangle srcRect,
-            GraphicsUnit srcUnit,
-            ImageAttributes imageAttr)
-        {
-            DrawImage(image, destPoints, srcRect, srcUnit, imageAttr, null, 0);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(
-            Image image,
-            Point[] destPoints,
-            Rectangle srcRect,
-            GraphicsUnit srcUnit,
-            ImageAttributes imageAttr,
-            DrawImageAbort callback)
-        {
-            DrawImage(image, destPoints, srcRect, srcUnit, imageAttr, callback, 0);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public unsafe void DrawImage(
-            Image image,
-            Point[] destPoints,
-            Rectangle srcRect,
-            GraphicsUnit srcUnit,
-            ImageAttributes imageAttr,
-            DrawImageAbort callback,
-            int callbackData)
-        {
-            if (destPoints == null)
-                throw new ArgumentNullException(nameof(destPoints));
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int count = destPoints.Length;
-            if (count != 3 && count != 4)
-                throw new ArgumentException(SR.GdiplusDestPointsInvalidLength);
-
-            fixed (Point* p = destPoints)
-            {
-                int status = Gdip.GdipDrawImagePointsRectI(
-                    new HandleRef(this, NativeGraphics),
-                    new HandleRef(image, image.nativeImage),
-                    p, destPoints.Length,
-                    srcRect.X, srcRect.Y, srcRect.Width, srcRect.Height,
-                    srcUnit,
-                    new HandleRef(imageAttr, imageAttr?.nativeImageAttributes ?? IntPtr.Zero),
-                    callback,
-                    (IntPtr)callbackData);
-
-                IgnoreMetafileErrors(image, ref status);
-                CheckErrorStatus(status);
-            }
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(
-            Image image,
-            Rectangle destRect,
-            float srcX,
-            float srcY,
-            float srcWidth,
-            float srcHeight,
-            GraphicsUnit srcUnit)
-        {
-            DrawImage(image, destRect, srcX, srcY, srcWidth, srcHeight, srcUnit, null);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(
-            Image image,
-            Rectangle destRect,
-            float srcX,
-            float srcY,
-            float srcWidth,
-            float srcHeight,
-            GraphicsUnit srcUnit,
-            ImageAttributes imageAttrs)
-        {
-            DrawImage(image, destRect, srcX, srcY, srcWidth, srcHeight, srcUnit, imageAttrs, null);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(
-            Image image,
-            Rectangle destRect,
-            float srcX,
-            float srcY,
-            float srcWidth,
-            float srcHeight,
-            GraphicsUnit srcUnit,
-            ImageAttributes imageAttrs,
-            DrawImageAbort callback)
-        {
-            DrawImage(image, destRect, srcX, srcY, srcWidth, srcHeight, srcUnit, imageAttrs, callback, IntPtr.Zero);
-        }
-
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(
-            Image image,
-            Rectangle destRect,
-            float srcX,
-            float srcY,
-            float srcWidth,
-            float srcHeight,
-            GraphicsUnit srcUnit,
-            ImageAttributes imageAttrs,
-            DrawImageAbort callback,
-            IntPtr callbackData)
-        {
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int status = Gdip.GdipDrawImageRectRect(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(image, image.nativeImage),
-                destRect.X, destRect.Y, destRect.Width, destRect.Height,
-                srcX, srcY, srcWidth, srcHeight,
-                srcUnit,
-                new HandleRef(imageAttrs, imageAttrs?.nativeImageAttributes ?? IntPtr.Zero),
-                callback,
-                callbackData);
-
-            IgnoreMetafileErrors(image, ref status);
-            CheckErrorStatus(status);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(
-            Image image,
-            Rectangle destRect,
-            int srcX,
-            int srcY,
-            int srcWidth,
-            int srcHeight,
-            GraphicsUnit srcUnit)
-        {
-            DrawImage(image, destRect, srcX, srcY, srcWidth, srcHeight, srcUnit, null);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(
-            Image image,
-            Rectangle destRect,
-            int srcX,
-            int srcY,
-            int srcWidth,
-            int srcHeight,
-            GraphicsUnit srcUnit,
-            ImageAttributes imageAttr)
-        {
-            DrawImage(image, destRect, srcX, srcY, srcWidth, srcHeight, srcUnit, imageAttr, null);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(
-            Image image,
-            Rectangle destRect,
-            int srcX,
-            int srcY,
-            int srcWidth,
-            int srcHeight,
-            GraphicsUnit srcUnit,
-            ImageAttributes imageAttr,
-            DrawImageAbort callback)
-        {
-            DrawImage(image, destRect, srcX, srcY, srcWidth, srcHeight, srcUnit, imageAttr, callback, IntPtr.Zero);
-        }
-
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
-        public void DrawImage(
-            Image image,
-            Rectangle destRect,
-            int srcX,
-            int srcY,
-            int srcWidth,
-            int srcHeight,
-            GraphicsUnit srcUnit,
-            ImageAttributes imageAttrs,
-            DrawImageAbort callback,
-            IntPtr callbackData)
-        {
-            if (image == null)
-                throw new ArgumentNullException(nameof(image));
-
-            int status = Gdip.GdipDrawImageRectRectI(
-                new HandleRef(this, NativeGraphics),
-                new HandleRef(image, image.nativeImage),
-                destRect.X, destRect.Y, destRect.Width, destRect.Height,
-                srcX, srcY, srcWidth, srcHeight,
-                srcUnit,
-                new HandleRef(imageAttrs, imageAttrs?.nativeImageAttributes ?? IntPtr.Zero),
-                callback,
-                callbackData);
-
-            IgnoreMetafileErrors(image, ref status);
-            CheckErrorStatus(status);
         }
 
         [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly")]
@@ -2850,53 +1223,6 @@ namespace System.Drawing
                 SafeNativeMethods.IntDeleteObject(new HandleRef(null, s_halftonePalette));
                 s_halftonePalette = IntPtr.Zero;
             }
-        }
-
-        /// <summary>
-        /// GDI+ will return a 'generic error' with specific win32 last error codes when
-        /// a terminal server session has been closed, minimized, etc... We don't want
-        /// to throw when this happens, so we'll guard against this by looking at the
-        /// 'last win32 error code' and checking to see if it is either 1) access denied
-        /// or 2) proc not found and then ignore it.
-        ///
-        /// The problem is that when you lock the machine, the secure desktop is enabled and
-        /// rendering fails which is expected (since the app doesn't have permission to draw
-        /// on the secure desktop). Not sure if there's anything you can do, short of catching
-        /// the desktop switch message and absorbing all the exceptions that get thrown while
-        /// it's the secure desktop.
-        /// </summary>
-        private void CheckErrorStatus(int status)
-        {
-            if (status == Gdip.Ok)
-                return;
-
-            // Generic error from GDI+ can be GenericError or Win32Error.
-            if (status == Gdip.GenericError || status == Gdip.Win32Error)
-            {
-                int error = Marshal.GetLastWin32Error();
-                if (error == SafeNativeMethods.ERROR_ACCESS_DENIED || error == SafeNativeMethods.ERROR_PROC_NOT_FOUND ||
-                        // Here, we'll check to see if we are in a terminal services session...
-                        (((UnsafeNativeMethods.GetSystemMetrics(NativeMethods.SM_REMOTESESSION) & 0x00000001) != 0) && (error == 0)))
-                {
-                    return;
-                }
-            }
-
-            // Legitimate error, throw our status exception.
-            throw Gdip.StatusException(status);
-        }
-
-        /// <summary>
-        /// GDI+ will return a 'generic error' when we attempt to draw an Emf
-        /// image with width/height == 1. Here, we will hack around this by
-        /// resetting the errorstatus. Note that we don't do simple arg checking
-        /// for height || width == 1 here because transforms can be applied to
-        /// the Graphics object making it difficult to identify this scenario.
-        /// </summary>
-        private void IgnoreMetafileErrors(Image image, ref int errorStatus)
-        {
-            if (errorStatus != Gdip.Ok && image.RawFormat.Equals(ImageFormat.Emf))
-                errorStatus = Gdip.Ok;
         }
     }
 }
